// src/config/form-types.ts
// These steps enums serve as keys for our step object within our state.
export enum STEPS {
  STEP_ONE = "step-one",
  STEP_TWO = "step-two",
}

// These fields enums serve as keys for our fields within our step object.
export enum FIELDS {
  RATING = "rating",
  TITLE = "title",
  DESCRIPTION = "description",
  IMAGE = "image",
  IMAGE_DESCRIPTION = "image-description",
}